Fixed update mechanic.

+++ src/jobs/time_entries_sync_job.ts
@@ -287,10 +287,31 @@ export class TimeEntriesSyncJob extends SyncJob {
         // loop through all STEOs (except that which TE is updated last)
         for (const serviceTimeEntryObjectWrapper of timeEntrySyncedObjectWrapper.serviceTimeEntryObjectWrappers) {
           if (serviceTimeEntryObjectWrapper !== lastUpdatedServiceTimeEntryObjectWrapper) {
-            // delete current TE from the service and continue to the scenario e)
-            await serviceTimeEntryObjectWrapper.syncedService.deleteTimeEntry(serviceTimeEntryObjectWrapper.serviceTimeEntryObject.id);
-            // set as undefined => use scenario e)
-            serviceTimeEntryObjectWrapper.timeEntry = undefined;
+
+            if (serviceTimeEntryObjectWrapper.timeEntry) {
+              console.log("B1 - update")
+
+              const index = timeEntrySyncedObjectWrapper.timeEntrySyncedObject.serviceTimeEntryObjects.indexOf(serviceTimeEntryObjectWrapper.serviceTimeEntryObject);
+              if (index == -1) {
+                // shouldn't happen but better safe
+                Sentry.captureMessage(`Object not found ${serviceTimeEntryObjectWrapper.serviceTimeEntryObject}`);
+              }
+              timeEntrySyncedObjectWrapper.timeEntrySyncedObject.serviceTimeEntryObjects.splice(index, 1);
+
+              serviceTimeEntryObjectWrapper.timeEntry = await this._updateTimeEntry(
+                  serviceTimeEntryObjectWrapper.serviceDefinition,
+                  lastUpdatedServiceTimeEntryObjectWrapper.timeEntry,
+                  serviceTimeEntryObjectWrapper.timeEntry,
+                  otherServicesMappingsObjects,
+                  timeEntrySyncedObjectWrapper.timeEntrySyncedObject,
+                  serviceTimeEntryObjectWrapper.serviceTimeEntryObject.isOrigin
+              );
+            } else {
+              // delete current TE from the service and continue to the scenario e)
+              await serviceTimeEntryObjectWrapper.syncedService.deleteTimeEntry(serviceTimeEntryObjectWrapper.serviceTimeEntryObject.id);
+              // set as undefined => use scenario e)
+              serviceTimeEntryObjectWrapper.timeEntry = undefined;
+            }
           }
         }
       } else {
@@ -385,7 +406,7 @@ export class TimeEntriesSyncJob extends SyncJob {
   private async updateJobLog(errors: Timer2TicketError[])
   {
     this._jobLog.errors.concat(errors)
-    let updated = await databaseService.updateJobLog(this._jobLog);
+    const updated = await databaseService.updateJobLog(this._jobLog);
     if (updated instanceof JobLog) {
       this._jobLog = updated;
     }
@@ -430,20 +451,76 @@ export class TimeEntriesSyncJob extends SyncJob {
       serviceObjectsMappings,
     );
 
+    return await this.handleCreatedTimeEntry(
+      createdTimeEntry,
+      timeEntrySyncedObject,
+      serviceDefinition.name,
+      shouldBeOrigin,
+      service.errors
+    )
+  }
+
+  private async handleCreatedTimeEntry(
+    createdTimeEntry: TimeEntry | null,
+    timeEntrySyncedObject: TimeEntrySyncedObject,
+    serviceDefinitionName: string,
+    shouldBeOrigin: boolean,
+    errors: Timer2TicketError[]
+  ) {
     if (createdTimeEntry) {
       // push newly created STEOs (with isOrigin: false)
       timeEntrySyncedObject.serviceTimeEntryObjects.push(
-        new ServiceTimeEntryObject(createdTimeEntry.id, serviceDefinition.name, shouldBeOrigin)
+        new ServiceTimeEntryObject(createdTimeEntry.id, serviceDefinitionName, shouldBeOrigin)
       );
 
       // lastly created -> update lastUpdate (every created TE will update lastUpdated, but the last created one will be permanent)
       this._updateTimeEntrySyncedObject(timeEntrySyncedObject, createdTimeEntry.lastUpdated, createdTimeEntry.start);
-
-      await this.updateJobLog(service.errors)
-      return createdTimeEntry;
     }
-    await this.updateJobLog(service.errors)
-    return undefined;
+    await this.updateJobLog(errors)
+    return createdTimeEntry ?? undefined;
+  }
+
+  /**
+   * Updates a time entry for given service based on given TE model.
+   * Also creates new STEO and pushes it to the given TESO.
+   * Updates given TESO's lastUpdated property to updated TE's date.
+   * Returns updated TE
+   *
+   * @param serviceDefinition
+   * @param updatedTimeEntry
+   * @param originalTimeEntry
+   * @param otherServicesMappingsObjects
+   * @param timeEntrySyncedObject
+   * @param shouldBeOrigin
+   */
+  private async _updateTimeEntry(
+      serviceDefinition: ServiceDefinition,
+      updatedTimeEntry: TimeEntry,
+      originalTimeEntry: TimeEntry,
+      otherServicesMappingsObjects: MappingsObject[],
+      timeEntrySyncedObject: TimeEntrySyncedObject,
+      shouldBeOrigin: boolean) {
+
+    const service = SyncedServiceCreator.create(serviceDefinition, this._user);
+    const serviceObjectsMappings = otherServicesMappingsObjects
+      .filter(mappingsObject => mappingsObject.service === serviceDefinition.name)
+      .map(mappingsObject => new ServiceObject(mappingsObject.id, mappingsObject.name, mappingsObject.type));
+
+    const createdTimeEntry = await service.updateTimeEntry(
+        updatedTimeEntry.durationInMilliseconds,
+        new Date(updatedTimeEntry.start),
+        updatedTimeEntry.text,
+        serviceObjectsMappings,
+        originalTimeEntry
+    );
+
+    return await this.handleCreatedTimeEntry(
+      createdTimeEntry,
+      timeEntrySyncedObject,
+      serviceDefinition.name,
+      shouldBeOrigin,
+      service.errors
+    )
   }
 
   /**
